+++ app/controllers/custom/proposals_controller.rb
@@ -2,6 +2,8 @@ require_dependency Rails.root.join("app", "controllers", "proposals_controller")
 
 class ProposalsController
 
+  include ProposalsHelper
+
   before_action :authenticate_user!, except: [:index, :show, :map, :summary, :json_data]
   before_action :process_tags, only: [:create, :update]
 
@@ -16,29 +18,6 @@ class ProposalsController
     @proposals_coordinates = all_proposal_map_locations
   end
   
-  def all_proposal_map_locations
-    ids = if params[:search]
-      Proposal.search(params[:search]).pluck(:id)
-    elsif params[:tags]
-      Proposal.not_archived.published.tagged_with(params[:tags].split(","), all: true)
-    else
-      Proposal.not_archived.published.pluck(:id)
-    end
-      MapLocation.where(proposal_id: ids).map(&:json_data)
-  end
-
-  def json_data
-    proposal = Proposal.find(params[:id])
-    data = {
-      proposal_id: proposal.id,
-      proposal_title: proposal.title
-    }.to_json
-
-    respond_to do |format|
-      format.json { render json: data }
-    end
-  end
-
   private
     def process_tags
       params[:proposal][:tag_list_categories].split(",").each do |t|
